refactor: extract run function
